Move utils to src/utils
