+++ src/alt/addons.js
@@ -1,13 +1,13 @@
 import Alt from './'
 
-import ActionListeners from '../../utils/ActionListeners'
-import AltManager from '../../utils/AltManager'
-import DispatcherRecorder from '../../utils/DispatcherRecorder'
+import ActionListeners from '../utils/ActionListeners'
+import AltManager from '../utils/AltManager'
+import DispatcherRecorder from '../utils/DispatcherRecorder'
 
-import atomicTransactions from '../../utils/atomicTransactions'
-import chromeDebug from '../../utils/chromeDebug'
-import makeFinalStore from '../../utils/makeFinalStore'
-import withAltContext from '../../utils/withAltContext'
+import atomicTransactions from '../utils/atomicTransactions'
+import chromeDebug from '../utils/chromeDebug'
+import makeFinalStore from '../utils/makeFinalStore'
+import withAltContext from '../utils/withAltContext'
 
 import AltContainer from '../../AltContainer'
