+++ src/utils/AltManager.js
