+++ test/listen-to-actions.js
@@ -1,7 +1,7 @@
 import { assert } from 'chai'
 import Alt from '../dist/alt-with-runtime'
 
-import ActionListeners from '../utils/ActionListeners'
+import ActionListeners from '../src/utils/ActionListeners'
 
 const alt = new Alt()
